Hide parent-cluster branch & maximise screen space
(1) hide the branch from a (hidden) parent to the common ancestor of
a (monophyletic) cluster. This involves adding a "stub" node as
auspice always displays the branch from a non-hidden node to a (parent)
hidden node.

(2) modify the num_date & div attributes for hidden nodes such that
they do not take up screen space in auspice. I.e. they are modified so
that they are equal to the minimum values of non-hidden nodes in the tree.